restructure tests structure, make separate test suite for each BasicService method
